+++ src/main/java/com/drmangotea/tfmg/content/machinery/vat/base/VatItem.java
@@ -1,7 +1,7 @@
 package com.drmangotea.tfmg.content.machinery.vat.base;
 
 import com.drmangotea.tfmg.registry.TFMGBlockEntities;
-import com.simibubi.create.api.connectivity.ConnectivityHandler;
+import com.drmangotea.tfmg.base.TFMGBlockConnectivityHandler;
 import com.simibubi.create.content.fluids.tank.FluidTankBlockEntity;
 import com.simibubi.create.foundation.block.IBE;
 import net.minecraft.core.BlockPos;
@@ -80,7 +80,7 @@ public class VatItem extends BlockItem {
 
         if (!VatBlock.isVat(placedOnState))
             return;
-        VatBlockEntity tankAt = ConnectivityHandler.partAt(TFMGBlockEntities.CHEMICAL_VAT.get(), world, placedOnPos);
+        VatBlockEntity tankAt = TFMGBlockConnectivityHandler.partAt(TFMGBlockEntities.CHEMICAL_VAT.get(), world, placedOnPos);
         if (tankAt == null)
             return;
         VatBlockEntity controllerTE =  tankAt.getControllerBE();
